refactor(site): move name validation to utils
