+++ site/src/util/formUtils.ts
@@ -1,5 +1,18 @@
 import { FormikContextType, FormikErrors, getIn } from "formik"
 import { ChangeEvent, ChangeEventHandler, FocusEventHandler, ReactNode } from "react"
+import * as Yup from "yup"
+
+export const Language = {
+  nameRequired: (name: string): string => {
+    return `Please enter a ${name.toLowerCase()}.`
+  },
+  nameInvalidChars: (name: string): string => {
+    return `${name} must start with a-Z or 0-9 and can contain a-Z, 0-9 or -`
+  },
+  nameTooLong: (name: string): string => {
+    return `${name} cannot be longer than 32 characters`
+  },
+}
 
 interface FormHelpers {
   name: string
@@ -38,3 +51,16 @@ export const onChangeTrimmed =
     event.target.value = event.target.value.trim()
     form.handleChange(event)
   }
+
+// REMARK: Keep in sync with coderd/httpapi/httpapi.go#L40
+const maxLenName = 32
+
+// REMARK: Keep in sync with coderd/httpapi/httpapi.go#L18
+const usernameRE = /^[a-zA-Z0-9]+(?:-[a-zA-Z0-9]+)*$/
+
+// REMARK: see #1756 for name/username semantics
+export const nameValidator = (name: string): Yup.StringSchema =>
+  Yup.string()
+    .required(Language.nameRequired(name))
+    .matches(usernameRE, Language.nameInvalidChars(name))
+    .max(maxLenName, Language.nameTooLong(name))
